--- a/devito/finite_differences/differentiable.py
+++ b/devito/finite_differences/differentiable.py
@@ -27,10 +27,6 @@ class Differentiable(sympy.Expr, Evaluable):
 
     _state = ('space_order', 'time_order', 'indices')
 
-    @property
-    def _sympify(self):
-        return self
-
     @cached_property
     def _functions(self):
         return frozenset().union(*[i._functions for i in self._args_diff])
@@ -123,6 +119,19 @@ class Differentiable(sympy.Expr, Evaluable):
             # Cartesian grid, do no waste time
             return self
         return self.func(*[getattr(a, '_eval_at', lambda x: a)(func) for a in self.args])
+
+    def _subs(self, old, new, **hints):
+        if old is self:
+            return new
+        if old is new:
+            return self
+        args = list(self.args)
+        for i, arg in enumerate(args):
+            try:
+                args[i] = arg._subs(old, new, **hints)
+            except AttributeError:
+                continue
+        return self.func(*args, evaluate=False)
 
     @property
     def _eval_deriv(self):
@@ -292,14 +301,7 @@ class DifferentiableOp(Differentiable):
         return self.func(*[getattr(a, 'subs', lambda x: a)(*args, **kwargs)
                            for a in self.args], evaluate=False)
 
-    def _subs(self, old, new, **hints):
-        args = list(self.args)
-        for i, arg in enumerate(args):
-            try:
-                args[i] = arg._subs(old, new, **hints)
-            except AttributeError:
-                continue
-        return self.func(*args, evaluate=False)
+    _subs = Differentiable._subs
 
     @property
     def _gather_for_diff(self):