--- a/devito/types/dense.py
+++ b/devito/types/dense.py
@@ -33,7 +33,7 @@ __all__ = ['Function', 'TimeFunction']
 RegionMeta = namedtuple('RegionMeta', 'offset size')
 
 
-class DiscreteFunction(AbstractFunction, ArgProvider):
+class DiscreteFunction(AbstractFunction, ArgProvider, Differentiable):
 
     """
     Tensor symbol representing a discrete function in symbolic equations.
@@ -101,6 +101,8 @@ class DiscreteFunction(AbstractFunction, ArgProvider):
         # The only possibility for two DiscreteFunctions to be considered equal
         # is that they are indeed the same exact object
         return self is other
+
+    _subs = Differentiable._subs
 
     __hash__ = AbstractFunction.__hash__  # Required since we're overriding __eq__
 
@@ -832,7 +834,7 @@ class DiscreteFunction(AbstractFunction, ArgProvider):
         ['grid', 'staggered', 'initializer']
 
 
-class Function(DiscreteFunction, Differentiable):
+class Function(DiscreteFunction):
 
     """
     Tensor symbol representing a discrete function in symbolic equations.
@@ -1327,7 +1329,7 @@ class TimeFunction(Function):
         i = int(self.time_order / 2) if self.time_order >= 2 else 1
         _t = self.dimensions[self._time_position]
 
-        return self.subs({_t: _t + i * _t.spacing})
+        return self._subs(_t, _t + i * _t.spacing)
 
     @property
     def backward(self):
@@ -1335,7 +1337,7 @@ class TimeFunction(Function):
         i = int(self.time_order / 2) if self.time_order >= 2 else 1
         _t = self.dimensions[self._time_position]
 
-        return self.subs({_t: _t - i * _t.spacing})
+        return self._subs(_t, _t - i * _t.spacing)
 
     @property
     def _time_size(self):