--- a/setup.cfg
+++ b/setup.cfg
@@ -3,6 +3,11 @@ max-line-length = 90
 ignore = F403,E226,E731,W503,F405,E722,E741,W504,W605
 exclude = .github
 
+[flake8_nb]
+max-line-length = 90
+ignore = F403,E226,E731,W503,F405,E722,E741,W504,W605
+exclude = .github
+
 [tool:pytest]
 python_files = test_*.py *_example.py example_*.py
 addopts = --durations=20 --maxfail=5