… examples/cfd/06_poisson.ipynb