Minor tweaks to improve the speed of `evaluate` that is quite slow going through all the nobs of sympy. Bypasses a lot of it via direct call to `_subs` and avoid very expensive `evalf` on expression while only needed on the coefficients.

timing improvement for 12th order 3D on my laptop for this branch

```
In [1]: %timeit u.laplace.laplace.evaluate                                                                                                                                           
1.04 s ± 71.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [2]: %timeit u.laplace.laplace.evaluate                                                                                                                                           
1.06 s ± 132 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
vs master:

```
In [1]: %timeit u.laplace.laplace.evaluate   
6.23 s ± 1.02 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [2]: %timeit u.laplace.laplace.evaluate                                                                                                                                           
7.16 s ± 2.23 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [3]: %timeit u.laplace.laplace.evaluate                                                                                                                                           
5.56 s ± 268 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

And aside, `setup.cfg` updated for `flake8-nb` if someone wants to play with it but not adding it to CI.