@mloubout 2020-06-23
>, I got a question: why do we override _subs instead of _eval_subs, which seems to be the stub we need here?

`aresame`  lot more expensive than the `old is new` in the one here.